refactor!(Vortex80AEsc): Update to proper naming convention
